Get osd key from ceph
There have been situations where ceph_bootstrap_osd_key disappears from
node attributes. In situations like this we can still get the bootstrap
key from ceph itself as long as the monitors are in quorum. This commit
also adds a message when stuck within the loop waiting for the quorum,
this prevents the appearance of a hung chef-client run.

Signed-Off-By: JuanJose 'JJ' Galvez <juanjose.galvez@inktank.com>